Fix apg_project float64 crash on devices without double support; cast slice index to int
`DiffusionTransformer.apg_project()` unconditionally casts to float64,
which crashes on accelerators without float64 support (e.g. MPS:
"Cannot convert a MPS Tensor to float64 dtype"). This is the default
path (apg_scale=1.0), so every generation fails on such devices.

The projection now runs in the highest precision the device supports:
float64 everywhere, float32 on devices whose tensor API rejects double.
CUDA/CPU behavior is unchanged, and the result is still cast back to
the original input dtype, so only the intermediate precision differs.

Also fix a device-independent bug in the Gradio interface:
`seconds_total` arrives as a Python float from a Slider component, so
`audio[:,:,:seconds_total*sample_rate]` raises
"TypeError: slice indices must be integers" on any device.

Fixes #261

+++ stable_audio_tools/interface/interfaces/diffusion_cond.py
@@ -315,7 +315,9 @@ def generate_cond(
 
     # Cut the extra silence off the end, if the user requested a smaller seconds_total
     if cut_to_seconds_total:
-        audio = audio[:,:,:seconds_total*sample_rate]
+        # seconds_total arrives as a Python float (e.g. from a Gradio Slider);
+        # slice indices must be integers. int() truncates toward zero.
+        audio = audio[:,:,:int(seconds_total*sample_rate)]
 
     # Encode the audio to WAV format
     audio = rearrange(audio, "b d n -> d (b n)")
